Remove it again.
The implementation is based around the notion of "power", the number of elements in the set on the left times the number on the right.
Then it uses this to break ties/cycles.

It works for Zarf's Junkyward-Wars inventory problem, where all that really matters is that it produces an order,
but for a general-purpose programming language I don't think it's a good idea because the transitive relation of power can give unpredictable results.
In particular there are a lot of ways to break cycles, and this just picks the highest power.